Python SDK first pass
This gets the basics of the SDK working.
Note that api_client.py is based on the
same file from the swagger-codegen templates,
but with significant cleanup. Just a few of the
items remaining:
* Define enums
* Add inheritence for volume attachments.
* Hook up logging to the config.
* Add/clean up documentation.
* Improve error handling. (I put a basic
service error in place that's thown when the
service returns an error, but in general the
error handling story needs more thought.)