Check in generated code
For both Ruby and Python we're plannig to switch
to checking in the generated code. This just removes
the gitignore and adds one test. I'll be putting up
a corresponding CR in oracle-swagger-maven-plugin
shortly that includes the generator and templates.
Both the Python SDK generator and the templates are
small modifications of the OOTB python generation
that comes with swagger-codegen.