Switch to requests library
Also, updated the signer to something closer to the
signging example code (as suggested by Joe in a previous
review) so that it can plug into requests auth.

I have not hooked up the ability to specify certs -
opened https://jira.aka.lgl.grungy.us/browse/DEX-139 to track
that.