Waiters and test for launch instance tutorial
Add's a simple form of waiters. The wait_until code itself
was ported almost line by line from the Ruby version. However,
the interface is a little different: Instead of holding onto
a closure as in Ruby, I've added a Request object to the
response that has all the info to remake the request.
wait_until is then added to the context instead of to the response
itself, which avoids needing to reference the ApiClient from
the response. It's functional but slighly akward as it is now.
There are a few options for how this could be improved, where
the one I'm currently leaning toward would be to add a
generalized getter that would map to the getters in each API,
and then incorporate that into wait_until such that you only
need an OCID and not a seperate get call. Anyway, something
that will need to be discussed. For now, I think this is a
reasonable first take that will get people unblocked.

Also, added a test that runs through most of the 'launch your first
instance' tutorial - minus attaching the volume since that needs
an additional change first. This test was also ported from Ruby.
Note that the file name intentionally does not start with "test_"
so that it isn't run when doing 'python -m unittest discover'. I'd like
to switch over to pytest which should provide better mechanisms
for controlling which tests are run, but this works for now.

Also hooks up super basic support for debug logging: If you set
config.debugging to true then you'll see debug level logs in the
console. There's a jira for the beta version to make this more
full featured.