Consume Casper spec
This adds the casper api spec in the same way
as other specs, with a basic set of integration
tests. Support for streaming and special
auth handling will come later.

The casper-api.yaml is currently up for review and pending
some important edits to help with the generation, so for
now I've copied over the yaml into a temp folder and
and am using that to generate. I'll delete that and switch
the pom file over once the real spec is ready.