Streaming support for Casper
Adds support for streaming of Casper objects.
Responses from get_object() will have a DataStream
object in the response.data property. DataStream
has very basic support for iterating over the content
or getting all the content at once, and also exposes
the raw urllib3.response.HTTPResponse. Decoding is
currently left up to the consumer.

Input to put_object is currently limited to an IOBase
(basically a stream) or a string.

Successfully tested upload and download of a 2.6 GB
file.