Add Docs, Update Casper
- Adds a rough first pass at creating docs using
sphinx. The conf.py and Makefiles under docs/ were
both created using the sphinx quickstart command, with
a small number of edits. There are a lot of commands and
commented out options in there that we don't need at the
moment, but it seems to be common to check this in for
future use and so that you can see the available options.
- Updates the casper.yaml to match the one that Pritesh
is using for Java. (With the one edit of making the content
length not required for put_object - that will need to be
handled later if Casper in fact adds this.)
- Generates a new set of APIs and Models, with a few minor changes
to help clean up the docs.
- Adds a Makefile for common commands.