Put metadata on objects
Enables putting metadata on objects. That's enough to support
the CLI for now - getting metadata in objects will still just
show that data in the headers.

Also ups the version (yes - there are a lot of places to change
it right now, something that we should eventually figure out how
to fix).