Send reqeust ID and user agent
- Sends user agent
- Sends request ID (commented out for now - waiting to hear
from Casper exactly what they will be doing for GA)
- Updates logging, using the same log_requests parameters
that is currently used in Ruby. This will allow CLI users
to see the request ID that is sent with their request, along
with other request headers.
- Gets rid of the context, and updates each
api to accept a config and create its own ApiClient.
This will allow the CLI to more easily update the config
based on arguments, and to create only the 1 API that
it needs. It's also much closer to what we're doing in Ruby
and Java. Note that this is not necessarily the right final design for the
SDK (the goal for now is to support the CLI), but I do think that
breaking up the context class is a step in the right direction.