Remove certifi requirement
As discussed a while back, we'll make this not
a requirement for the SDK (but requests will use it
if it happens to be installed). We will keep it as a
requiremtn for the CLI (and I'll put up a separate CR
for the CLI to remove the version no certifi).