Move path expansion to signer, base_client defensively validates config
Pushing the expansion into the signer removes the need for callers to come through the config
module just to create a signer.  It may be the path we take for building the default clients,
but there's no need to force the coupling.

Since the client always needs valid config, it's easier to always validate at the point
of usage, instead of relying on users to validate config on their own (and getting cryptic
messages when they fail to validate).