Split up generated code by spec
This is the same structure used by Ruby, and similar
to what is used by Java. Tests pass. Note that a type
mapping now must be passed into BaseClient in order
for it to resolve the types for a particular service.

If we take this change we'll also need to update example
code and the changelog, but first I want to get agreement
on the code change itself.