Enable test runs in Team City
Follows the same pattern as Ruby and the CLI: checks
in the keys used for the tests, where the pass phrase
must be set in an environment variable.
Also adds the integ tests to the tox runs. (The large file
test had been failing in python 2 during a check to ensure
that the max memory usage was less than the file size. It turned
out that the max memory usage slightly exceeded the file size
even before performing any operations. The solution was just to
increase the file size.)