Allow for explicit setting of null/None values
Uses the Sentinel class (reading the comments this seems to be the intent). Note that if an explicit null comes back from the service, now the field
will be the None Sentinel. This means that "if <value>" will be false (as the None Sentinel is falsey) but "if <val> is None" will be false because
the sentinel is not the NoneType