Add function to check if a value is in an allowed set or an acceptable None (literal None or the None Sentinel)
This is intended for enum values as currently it's not possible to set an enum to None. It will either exception out (if it is part of a request) or
if it is in the response a None value will be translated to UNKNOWN_ENUM_VALUE (which, strictly speaking, isn't correct and could give callers the wrong
idea).

There'll also need to be a codegen change to leverage this function