Add .readthedocs.yml config file
This file tells ReadtheDocs to not build any of the optional formats.
The readthedocssinglehtmllocalmedia build was taking too long and
using too much memory causing the builds to fail.

This config file was tested by forking the GitHub repo and then
creating a new ReadtheDocs project on the fork.  The builds failed
as expected.  The configuration file was added to the fork and
the ReadtheDocs build passed in 618 seconds.