Fix test_basic_wait
It was possible for the first get call inside tge 'with' block to
throw an exception which caused response to not get updated.  When
The wait_until call was made later it was possible for the response
to still be from the delete call which would make wait_until fail.