Update test_waiters to use unique names
Test waiters does not use VCR, but the utility that creates the
unique names does, and in the case where it detects VCR it just appends
"_vcr" to the name.  This updates test_waiters to tell unique_name
to ignore_vcr and produce a unique name.

This fix elimiates one area where multiple builds were conflicting when
trying to use the same resources.