Fix object-storage test which are failing due to missing resources
Split upload manager tests into their own test file
Fix set_up_test_data so that it is required by tests which need certain buckets
Fix scope of set_up_test_data and it's dependancies
Add object-storage tests into the pr_build script
Re-record new cassettes
Removed cassettes that are no longer referenced.
Removed buckets which were not used from the tenancy and the function
ensure_test_data

set_up_test_data was not running because none of the tests had a dependency on
it.  It also had a problem with the scope.  It was set to module but namespace
and object_storage were both at function level scope.