Fix list_suppressions_with_sorting in test_email
The Python sort algorithm and the sort algorithm used by the email service
do not agree on how to handle underscores.  This caused the test
to fail when email accounts with leading underscores were added to
the email suppression list.

Remove all email addresses that were not created as part of the test
from the returned results and make sure that the remaining address
are sorted in the expected order.

Verified with python 2.7.12 and 3.5.1 with and without recordings.