Update pom.xml for DEXREQ-94
The changes for DEXREQ-94 were added with https://bitbucket.oci.oraclecorp.com/projects/SDK/repos/python-sdk/pull-requests/788/overview.
However with the confussion surrounding the KMS changes, it looks like the pom was not updated.
This updates the casper spec version 1.1.22 as specified in DEXREQ-94.  The SDK was then regenerated
and no additional changes occurred.