Skipping tagging tests when not using recorded responses
Tags can not be deleted and our tag namespace has hit the limit
for new tags.  The tests using recorded responses continue to work
but the tests fail when run in the "OCI: Run tests and re-record
output" teamcity job.  For now just skip the tagging tests when
re-recording output.