Update generate_service_rst_files.py
The client docs for Key Management and Compute Management were not
being generated.  The docs for the models were generated and on
ReadtheDocs, but the client docs were missing.  There is an
additional location in generate_service_rst_files.py which needs to
be updated to get the clients to generate and appear in the API
landing page.