Fix a bug in generate_service_rst_file.py
In Python 2 inspect.getfile() will return the *.pyc file instead of
the *.py file if the *.pyc file exists.  When first install oci
with `pip install -e .` When I initially tested this program
the .pyc files did not exist.  Once oci is imported into a
program it the .pyc files are generated, so the second run of
generate_service_rst_file.py would fail.