Add example showing how to use work requests when launching an instance
This commit adds an example illustrating how to use work requests with
instance launching. The example will launch an instance and wait for the
launch's work request to reach a terminal state (FAILED or SUCCEEDED).

It will then print out details about the work request (the result of a
GET for the work request, as well as any error and informational log
messages) and the launched instance. It will then terminate the
instance.