Fixes a bug where the source region is always picked up from the config file.
While creating the source region ObjectStorageClient, the current code uses the region which is set in the `config` object which was created by reading the config file(which may have a different region set). The log lines print the source_region variable but never set it inside the config object.